add: new migration adjusted

+++ database/migrations/2026_07_14_090121_add_two_factor_enabled_to_users_table.php
@@ -6,23 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->boolean('two_factor_enabled')->default(false)->after('status');
-        });
+        if (Schema::hasTable('users') && !Schema::hasColumn('users', 'two_factor_enabled')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->boolean('two_factor_enabled')->default(false)->after('role');
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('two_factor_enabled');
-        });
+        if (Schema::hasTable('users') && Schema::hasColumn('users', 'two_factor_enabled')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('two_factor_enabled');
+            });
+        }
     }
 };
